lcas2/requirements.txt
# LCAS v4.0 - Complete Requirements
# Core dependencies for Legal Case Analysis System

# Core Python packages
numpy>=1.24.0
pandas>=2.0.0
python-dateutil>=2.8.2
pyyaml>=6.0.1
python-dotenv>=1.0.0
tqdm>=4.66.0

# AI Providers (Optional but recommended)
openai>=1.0.0
anthropic>=0.21.0
httpx>=0.25.0
requests>=2.31.0

# File Processing
PyPDF2>=3.0.1
pdfplumber>=0.9.0
python-docx>=0.8.11
openpyxl>=3.1.2
chardet>=5.2.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# NLP and Text Processing
nltk>=3.8.1
textblob>=0.17.1

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# GUI Framework
tkinter-tooltip>=2.0.0

# Utilities
pathlib2>=2.3.7
click>=8.1.0
rich>=13.0.0

# Optional: Advanced Features
# Uncomment these for enhanced functionality:
# sentence-transformers>=2.2.2
# spacy>=3.7.0
# scikit-learn>=1.3.0
# faiss-cpu>=1.7.4
# transformers>=4.30.0
# customtkinter>=5.2.0
# pillow>=10.0.0